Move unsafe code from Nokhwa implementation to FFI layer
